Network.AWS.DAX

Description

DAX is a managed caching service engineered for Amazon DynamoDB. DAX dramatically speeds up database reads by caching frequently-accessed data from DynamoDB, so applications can access that data with sub-millisecond latency. You can create a DAX cluster easily, using the AWS Management Console. With a few simple modifications to your code, your application can begin taking advantage of the DAX cluster and realize significant improvements in read performance.

Service Configuration

dax :: Service

API version 2017-04-19 of the Amazon DynamoDB Accelerator (DAX) SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by DAX.

SubnetQuotaExceededFault

_SubnetQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

The request cannot be processed because it would exceed the allowed number of subnets in a subnet group.

ParameterGroupNotFoundFault

_ParameterGroupNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified parameter group does not exist.

InvalidParameterGroupStateFault

_InvalidParameterGroupStateFault :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters in a parameter group are in an invalid state.

SubnetGroupInUseFault

_SubnetGroupInUseFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified subnet group is currently in use.

ParameterGroupAlreadyExistsFault

_ParameterGroupAlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified parameter group already exists.

InvalidSubnet

_InvalidSubnet :: AsError a => Getting (First ServiceError) a ServiceError

An invalid subnet identifier was specified.

TagQuotaPerResourceExceeded

_TagQuotaPerResourceExceeded :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded the maximum number of tags for this DAX cluster.

ClusterNotFoundFault

_ClusterNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The requested cluster ID does not refer to an existing DAX cluster.

TagNotFoundFault

_TagNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The tag does not exist.

NodeQuotaForClusterExceededFault

_NodeQuotaForClusterExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of nodes for a DAX cluster.

InvalidClusterStateFault

_InvalidClusterStateFault :: AsError a => Getting (First ServiceError) a ServiceError

The requested DAX cluster is not in the available state.

InsufficientClusterCapacityFault

_InsufficientClusterCapacityFault :: AsError a => Getting (First ServiceError) a ServiceError

There are not enough system resources to create the cluster you requested (or to resize an already-existing cluster).

NodeNotFoundFault

_NodeNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

None of the nodes in the cluster have the given node ID.

ParameterGroupQuotaExceededFault

_ParameterGroupQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of parameter groups.

InvalidParameterValueException

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

The value for a parameter is invalid.

InvalidVPCNetworkStateFault

_InvalidVPCNetworkStateFault :: AsError a => Getting (First ServiceError) a ServiceError

The VPC network is in an invalid state.

SubnetInUse

_SubnetInUse :: AsError a => Getting (First ServiceError) a ServiceError

The requested subnet is being used by another subnet group.

ClusterQuotaForCustomerExceededFault

_ClusterQuotaForCustomerExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of DAX clusters for your AWS account.

SubnetGroupNotFoundFault

_SubnetGroupNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The requested subnet group name does not refer to an existing subnet group.

SubnetGroupAlreadyExistsFault

_SubnetGroupAlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

The specified subnet group already exists.

NodeQuotaForCustomerExceededFault

_NodeQuotaForCustomerExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have attempted to exceed the maximum number of nodes for your AWS account.

SubnetGroupQuotaExceededFault

_SubnetGroupQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

The request cannot be processed because it would exceed the allowed number of subnets in a subnet group.

ClusterAlreadyExistsFault

_ClusterAlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

You already have a DAX cluster with the given identifier.

InvalidARNFault

_InvalidARNFault :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Resource Name (ARN) supplied in the request is not valid.

InvalidParameterCombinationException

_InvalidParameterCombinationException :: AsError a => Getting (First ServiceError) a ServiceError

Two or more incompatible parameters were specified.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeClusters

module Network.AWS.DAX.DescribeClusters

DescribeParameters

module Network.AWS.DAX.DescribeParameters

DescribeEvents

module Network.AWS.DAX.DescribeEvents

IncreaseReplicationFactor

module Network.AWS.DAX.IncreaseReplicationFactor

CreateSubnetGroup

module Network.AWS.DAX.CreateSubnetGroup

DeleteCluster

module Network.AWS.DAX.DeleteCluster

UpdateCluster

module Network.AWS.DAX.UpdateCluster

CreateCluster

module Network.AWS.DAX.CreateCluster

DescribeDefaultParameters

module Network.AWS.DAX.DescribeDefaultParameters

DeleteParameterGroup

module Network.AWS.DAX.DeleteParameterGroup

UpdateParameterGroup

module Network.AWS.DAX.UpdateParameterGroup

DescribeSubnetGroups

module Network.AWS.DAX.DescribeSubnetGroups

CreateParameterGroup

module Network.AWS.DAX.CreateParameterGroup

UpdateSubnetGroup

module Network.AWS.DAX.UpdateSubnetGroup

DeleteSubnetGroup

module Network.AWS.DAX.DeleteSubnetGroup

DescribeParameterGroups

module Network.AWS.DAX.DescribeParameterGroups

TagResource

module Network.AWS.DAX.TagResource

ListTags

module Network.AWS.DAX.ListTags

DecreaseReplicationFactor

module Network.AWS.DAX.DecreaseReplicationFactor

UntagResource

module Network.AWS.DAX.UntagResource

RebootNode

module Network.AWS.DAX.RebootNode

Types

ChangeType

data ChangeType

Constructors

Immediate

RequiresReboot

IsModifiable

data IsModifiable

Constructors

Conditional

False'

True'

ParameterType

data ParameterType

Constructors

Default

NodeTypeSpecific

SourceType

data SourceType

Constructors

Cluster

ParameterGroup

SubnetGroup

Cluster

data Cluster

Contains all of the attributes of a specific DAX cluster.

See: cluster smart constructor.

cluster :: Cluster

Creates a value of Cluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cStatus - The current status of the cluster.

cIAMRoleARN - A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

cClusterARN - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

cActiveNodes - The number of nodes in the cluster that are active (i.e., capable of serving requests).

cSecurityGroups - A list of security groups, and the status of each, for the nodes in the cluster.

cNotificationConfiguration - Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

cNodeIdsToRemove - A list of nodes to be removed from the cluster.

cTotalNodes - The total number of nodes in the cluster.

cPreferredMaintenanceWindow - A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00 . Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

cSubnetGroup - The subnet group where the DAX cluster is running.

cClusterName - The name of the DAX cluster.

cNodeType - The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

cNodes - A list of nodes that are currently in the cluster.

cClusterDiscoveryEndpoint - The configuration endpoint for this DAX cluster, consisting of a DNS name and a port number. Client applications can specify this endpoint, rather than an individual node endpoint, and allow the DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

cDescription - The description of the cluster.

cParameterGroup - The parameter group being used by nodes in the cluster.

cStatus :: Lens' Cluster (Maybe Text)

The current status of the cluster.

cIAMRoleARN :: Lens' Cluster (Maybe Text)

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

cClusterARN :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

cActiveNodes :: Lens' Cluster (Maybe Int)

The number of nodes in the cluster that are active (i.e., capable of serving requests).

cSecurityGroups :: Lens' Cluster [SecurityGroupMembership]

A list of security groups, and the status of each, for the nodes in the cluster.

cNotificationConfiguration :: Lens' Cluster (Maybe NotificationConfiguration)

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

cNodeIdsToRemove :: Lens' Cluster [Text]

A list of nodes to be removed from the cluster.

cTotalNodes :: Lens' Cluster (Maybe Int)

The total number of nodes in the cluster.

cPreferredMaintenanceWindow :: Lens' Cluster (Maybe Text)

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00 . Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

cSubnetGroup :: Lens' Cluster (Maybe Text)

The subnet group where the DAX cluster is running.

cClusterName :: Lens' Cluster (Maybe Text)

The name of the DAX cluster.

cNodeType :: Lens' Cluster (Maybe Text)

The node type for the nodes in the cluster. (All nodes in a DAX cluster are of the same type.)

cNodes :: Lens' Cluster [Node]

A list of nodes that are currently in the cluster.

cClusterDiscoveryEndpoint :: Lens' Cluster (Maybe Endpoint)

The configuration endpoint for this DAX cluster, consisting of a DNS name and a port number. Client applications can specify this endpoint, rather than an individual node endpoint, and allow the DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

cDescription :: Lens' Cluster (Maybe Text)

The description of the cluster.

cParameterGroup :: Lens' Cluster (Maybe ParameterGroupStatus)

The parameter group being used by nodes in the cluster.

Endpoint

data Endpoint

Represents the information required for client programs to connect to the configuration endpoint for a DAX cluster, or to an individual node within the cluster.

See: endpoint smart constructor.

endpoint :: Endpoint

Creates a value of Endpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eAddress - The DNS hostname of the endpoint.

ePort - The port number that applications should use to connect to the endpoint.

eAddress :: Lens' Endpoint (Maybe Text)

The DNS hostname of the endpoint.

ePort :: Lens' Endpoint (Maybe Int)

The port number that applications should use to connect to the endpoint.

Event

data Event

Represents a single occurrence of something interesting within the system. Some examples of events are creating a DAX cluster, adding or removing a node, or rebooting a node.

See: event smart constructor.

event :: Event

Creates a value of Event with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eSourceName - The source of the event. For example, if the event occurred at the node level, the source would be the node ID.

eSourceType - Specifies the origin of this event - a cluster, a parameter group, a node ID, etc.

eDate - The date and time when the event occurred.

eMessage - A user-defined message associated with the event.

eSourceName :: Lens' Event (Maybe Text)

The source of the event. For example, if the event occurred at the node level, the source would be the node ID.

eSourceType :: Lens' Event (Maybe SourceType)

Specifies the origin of this event - a cluster, a parameter group, a node ID, etc.

eDate :: Lens' Event (Maybe UTCTime)

The date and time when the event occurred.

eMessage :: Lens' Event (Maybe Text)

A user-defined message associated with the event.

Node

data Node

Represents an individual node within a DAX cluster.

See: node smart constructor.

node :: Node

Creates a value of Node with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nNodeStatus - The current status of the node. For example: available .

nParameterGroupStatus - The status of the parameter group associated with this node. For example, in-sync .

nAvailabilityZone - The Availability Zone (AZ) in which the node has been deployed.

nNodeId - A system-generated identifier for the node.

nEndpoint - The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

nNodeCreateTime - The date and time (in UNIX epoch format) when the node was launched.

nNodeStatus :: Lens' Node (Maybe Text)

The current status of the node. For example: available .

nParameterGroupStatus :: Lens' Node (Maybe Text)

The status of the parameter group associated with this node. For example, in-sync .

nAvailabilityZone :: Lens' Node (Maybe Text)

The Availability Zone (AZ) in which the node has been deployed.

nNodeId :: Lens' Node (Maybe Text)

A system-generated identifier for the node.

nEndpoint :: Lens' Node (Maybe Endpoint)

The endpoint for the node, consisting of a DNS name and a port number. Client applications can connect directly to a node endpoint, if desired (as an alternative to allowing DAX client software to intelligently route requests and responses to nodes in the DAX cluster.

nNodeCreateTime :: Lens' Node (Maybe UTCTime)

The date and time (in UNIX epoch format) when the node was launched.

NodeTypeSpecificValue

data NodeTypeSpecificValue

Represents a parameter value that is applicable to a particular node type.

See: nodeTypeSpecificValue smart constructor.

nodeTypeSpecificValue :: NodeTypeSpecificValue

Creates a value of NodeTypeSpecificValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ntsvValue - The parameter value for this node type.

ntsvNodeType - A node type to which the parameter value applies.

ntsvValue :: Lens' NodeTypeSpecificValue (Maybe Text)

The parameter value for this node type.

ntsvNodeType :: Lens' NodeTypeSpecificValue (Maybe Text)

A node type to which the parameter value applies.

NotificationConfiguration

data NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for publishing DAX events to subscribers using Amazon Simple Notification Service (SNS).

See: notificationConfiguration smart constructor.

notificationConfiguration :: NotificationConfiguration

Creates a value of NotificationConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ncTopicStatus - The current state of the topic.

ncTopicARN - The Amazon Resource Name (ARN) that identifies the topic.

ncTopicStatus :: Lens' NotificationConfiguration (Maybe Text)

The current state of the topic.

ncTopicARN :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

Parameter

data Parameter

Describes an individual setting that controls some aspect of DAX behavior.

See: parameter smart constructor.

parameter :: Parameter

Creates a value of Parameter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pParameterValue - The value for the parameter.

pParameterType - Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

pSource - How the parameter is defined. For example, system denotes a system-defined parameter.

pIsModifiable - Whether the customer is allowed to modify the parameter.

pDataType - The data type of the parameter. For example, integer :

pNodeTypeSpecificValues - A list of node types, and specific parameter values for each node.

pAllowedValues - A range of values within which the parameter can be set.

pParameterName - The name of the parameter.

pDescription - A description of the parameter

pChangeType - The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

pParameterValue :: Lens' Parameter (Maybe Text)

The value for the parameter.

pParameterType :: Lens' Parameter (Maybe ParameterType)

Determines whether the parameter can be applied to any nodes, or only nodes of a particular type.

pSource :: Lens' Parameter (Maybe Text)

How the parameter is defined. For example, system denotes a system-defined parameter.

pIsModifiable :: Lens' Parameter (Maybe IsModifiable)

Whether the customer is allowed to modify the parameter.

pDataType :: Lens' Parameter (Maybe Text)

The data type of the parameter. For example, integer :

pNodeTypeSpecificValues :: Lens' Parameter [NodeTypeSpecificValue]

A list of node types, and specific parameter values for each node.

pAllowedValues :: Lens' Parameter (Maybe Text)

A range of values within which the parameter can be set.

pParameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

pDescription :: Lens' Parameter (Maybe Text)

A description of the parameter

pChangeType :: Lens' Parameter (Maybe ChangeType)

The conditions under which changes to this parameter can be applied. For example, requires-reboot indicates that a new value for this parameter will only take effect if a node is rebooted.

ParameterGroup

data ParameterGroup

A named set of parameters that are applied to all of the nodes in a DAX cluster.

See: parameterGroup smart constructor.

parameterGroup :: ParameterGroup

Creates a value of ParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pgDescription - A description of the parameter group.

pgParameterGroupName - The name of the parameter group.

pgDescription :: Lens' ParameterGroup (Maybe Text)

A description of the parameter group.

pgParameterGroupName :: Lens' ParameterGroup (Maybe Text)

The name of the parameter group.

ParameterGroupStatus

data ParameterGroupStatus

The status of a parameter group.

See: parameterGroupStatus smart constructor.

parameterGroupStatus :: ParameterGroupStatus

Creates a value of ParameterGroupStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pgsNodeIdsToReboot - The node IDs of one or more nodes to be rebooted.

pgsParameterApplyStatus - The status of parameter updates.

pgsParameterGroupName - The name of the parameter group.

pgsNodeIdsToReboot :: Lens' ParameterGroupStatus [Text]

The node IDs of one or more nodes to be rebooted.

pgsParameterApplyStatus :: Lens' ParameterGroupStatus (Maybe Text)

The status of parameter updates.

pgsParameterGroupName :: Lens' ParameterGroupStatus (Maybe Text)

The name of the parameter group.

ParameterNameValue

data ParameterNameValue

An individual DAX parameter.

See: parameterNameValue smart constructor.

parameterNameValue :: ParameterNameValue

Creates a value of ParameterNameValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pnvParameterValue - The value of the parameter.

pnvParameterName - The name of the parameter.

pnvParameterValue :: Lens' ParameterNameValue (Maybe Text)

The value of the parameter.

pnvParameterName :: Lens' ParameterNameValue (Maybe Text)

The name of the parameter.

SecurityGroupMembership

data SecurityGroupMembership

An individual VPC security group and its status.

See: securityGroupMembership smart constructor.

securityGroupMembership :: SecurityGroupMembership

Creates a value of SecurityGroupMembership with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sgmStatus - The status of this security group.

sgmSecurityGroupIdentifier - The unique ID for this security group.

sgmStatus :: Lens' SecurityGroupMembership (Maybe Text)

The status of this security group.

sgmSecurityGroupIdentifier :: Lens' SecurityGroupMembership (Maybe Text)

The unique ID for this security group.

Subnet

data Subnet

Represents the subnet associated with a DAX cluster. This parameter refers to subnets defined in Amazon Virtual Private Cloud (Amazon VPC) and used with DAX.

See: subnet smart constructor.

subnet :: Subnet

Creates a value of Subnet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sSubnetIdentifier - The system-assigned identifier for the subnet.

sSubnetAvailabilityZone - The Availability Zone (AZ) for subnet subnet.

sSubnetIdentifier :: Lens' Subnet (Maybe Text)

The system-assigned identifier for the subnet.

sSubnetAvailabilityZone :: Lens' Subnet (Maybe Text)

The Availability Zone (AZ) for subnet subnet.

SubnetGroup

data SubnetGroup

Represents the output of one of the following actions:

CreateSubnetGroup

ModifySubnetGroup

See: subnetGroup smart constructor.

subnetGroup :: SubnetGroup

Creates a value of SubnetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sgVPCId - The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

sgSubnets - A list of subnets associated with the subnet group.

sgSubnetGroupName - The name of the subnet group.

sgDescription - The description of the subnet group.

sgVPCId :: Lens' SubnetGroup (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

sgSubnets :: Lens' SubnetGroup [Subnet]

A list of subnets associated with the subnet group.

sgSubnetGroupName :: Lens' SubnetGroup (Maybe Text)

The name of the subnet group.

sgDescription :: Lens' SubnetGroup (Maybe Text)

The description of the subnet group.

Tag

data Tag

A description of a tag. Every tag is a key-value pair. You can add up to 50 tags to a single DAX cluster.

AWS-assigned tag names and values are automatically assigned the aws: prefix, which the user cannot assign. AWS-assigned tag names do not count towards the tag limit of 50. User-assigned tag names have the prefix user: .

You cannot backdate the application of a tag.

See: tag smart constructor.

tag :: Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value of the tag. Tag values are case-sensitive and can be null.

tagKey - The key for the tag. Tag keys are case sensitive. Every DAX cluster can only have one tag with the same key. If you try to add an existing tag (same key), the existing tag value will be updated to the new value.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag. Tag values are case-sensitive and can be null.

tagKey :: Lens' Tag (Maybe Text)

The key for the tag. Tag keys are case sensitive. Every DAX cluster can only have one tag with the same key. If you try to add an existing tag (same key), the existing tag value will be updated to the new value.