Network.AWS.DAX.CreateCluster

Description

Creates a DAX cluster. All nodes in the cluster run the same DAX caching software.

Creating a Request

createCluster

Arguments

:: Text
ccClusterName

-> Text
ccNodeType

-> Int
ccReplicationFactor

-> Text
ccIAMRoleARN

-> CreateCluster

Creates a value of CreateCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccSecurityGroupIds - A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.) If this parameter is not specified, DAX assigns the default VPC security group to each node.

ccSubnetGroupName - The name of the subnet group to be used for the replication group. Important: DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

ccPreferredMaintenanceWindow - Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:05:00-sun:09:00

ccAvailabilityZones - The Availability Zones (AZs) in which the cluster nodes will be created. All nodes belonging to the cluster are placed in these Availability Zones. Use this parameter if you want to distribute the nodes across multiple AZs.

ccDescription - A description of the cluster.

ccNotificationTopicARN - The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

ccTags - A set of tags to associate with the DAX cluster.

ccParameterGroupName - The parameter group to be associated with the DAX cluster.

ccClusterName - The cluster identifier. This parameter is stored as a lowercase string. Constraints: * A name must contain from 1 to 20 alphanumeric characters or hyphens. * The first character must be a letter. * A name cannot end with a hyphen or contain two consecutive hyphens.

ccNodeType - The compute and memory capacity of the nodes in the cluster.

ccReplicationFactor - The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to 2 or more.

ccIAMRoleARN - A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

data CreateCluster

See: createCluster smart constructor.

Request Lenses

ccSecurityGroupIds :: Lens' CreateCluster [Text]

A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.) If this parameter is not specified, DAX assigns the default VPC security group to each node.

ccSubnetGroupName :: Lens' CreateCluster (Maybe Text)

The name of the subnet group to be used for the replication group. Important: DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

ccPreferredMaintenanceWindow :: Lens' CreateCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are: * sun * mon * tue * wed * thu * fri * sat Example: sun:05:00-sun:09:00

ccAvailabilityZones :: Lens' CreateCluster [Text]

The Availability Zones (AZs) in which the cluster nodes will be created. All nodes belonging to the cluster are placed in these Availability Zones. Use this parameter if you want to distribute the nodes across multiple AZs.

ccDescription :: Lens' CreateCluster (Maybe Text)

A description of the cluster.

ccNotificationTopicARN :: Lens' CreateCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

ccTags :: Lens' CreateCluster [Tag]

A set of tags to associate with the DAX cluster.

ccParameterGroupName :: Lens' CreateCluster (Maybe Text)

The parameter group to be associated with the DAX cluster.

ccClusterName :: Lens' CreateCluster Text

The cluster identifier. This parameter is stored as a lowercase string. Constraints: * A name must contain from 1 to 20 alphanumeric characters or hyphens. * The first character must be a letter. * A name cannot end with a hyphen or contain two consecutive hyphens.

ccNodeType :: Lens' CreateCluster Text

The compute and memory capacity of the nodes in the cluster.

ccReplicationFactor :: Lens' CreateCluster Int

The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to 2 or more.

ccIAMRoleARN :: Lens' CreateCluster Text

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

Destructuring the Response

createClusterResponse

Arguments

:: Int
ccrsResponseStatus

-> CreateClusterResponse

Creates a value of CreateClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccrsCluster - A description of the DAX cluster that you have created.

ccrsResponseStatus - -- | The response status code.

data CreateClusterResponse

See: createClusterResponse smart constructor.

Response Lenses

ccrsCluster :: Lens' CreateClusterResponse (Maybe Cluster)

A description of the DAX cluster that you have created.

ccrsResponseStatus :: Lens' CreateClusterResponse Int

- | The response status code.