Network.AWS.DAX.CreateSubnetGroup

Description

Creates a new subnet group.

Creating a Request

createSubnetGroup

Arguments

:: Text
csgSubnetGroupName

-> CreateSubnetGroup

Creates a value of CreateSubnetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csgDescription - A description for the subnet group

csgSubnetGroupName - A name for the subnet group. This value is stored as a lowercase string.

csgSubnetIds - A list of VPC subnet IDs for the subnet group.

data CreateSubnetGroup

See: createSubnetGroup smart constructor.

Request Lenses

csgDescription :: Lens' CreateSubnetGroup (Maybe Text)

A description for the subnet group

csgSubnetGroupName :: Lens' CreateSubnetGroup Text

A name for the subnet group. This value is stored as a lowercase string.

csgSubnetIds :: Lens' CreateSubnetGroup [Text]

A list of VPC subnet IDs for the subnet group.

Destructuring the Response

createSubnetGroupResponse

Arguments

:: Int
csgrsResponseStatus

-> CreateSubnetGroupResponse

Creates a value of CreateSubnetGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csgrsSubnetGroup - Represents the output of a CreateSubnetGroup operation.

csgrsResponseStatus - -- | The response status code.

data CreateSubnetGroupResponse

See: createSubnetGroupResponse smart constructor.

Response Lenses

csgrsSubnetGroup :: Lens' CreateSubnetGroupResponse (Maybe SubnetGroup)

Represents the output of a CreateSubnetGroup operation.

csgrsResponseStatus :: Lens' CreateSubnetGroupResponse Int

- | The response status code.