Network.AWS.DAX.DescribeDefaultParameters

Description

Returns the default system parameter information for the DAX caching software.

Creating a Request

describeDefaultParameters :: DescribeDefaultParameters

Creates a value of DescribeDefaultParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddpNextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults .

ddpMaxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved. The value for MaxResults must be between 20 and 100.

data DescribeDefaultParameters

See: describeDefaultParameters smart constructor.

Request Lenses

ddpNextToken :: Lens' DescribeDefaultParameters (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults .

ddpMaxResults :: Lens' DescribeDefaultParameters (Maybe Int)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved. The value for MaxResults must be between 20 and 100.

Destructuring the Response

describeDefaultParametersResponse

Arguments

:: Int
ddprsResponseStatus

-> DescribeDefaultParametersResponse

Creates a value of DescribeDefaultParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddprsNextToken - Provides an identifier to allow retrieval of paginated results.

ddprsParameters - A list of parameters. Each element in the list represents one parameter.

ddprsResponseStatus - -- | The response status code.

data DescribeDefaultParametersResponse

See: describeDefaultParametersResponse smart constructor.

Response Lenses

ddprsNextToken :: Lens' DescribeDefaultParametersResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

ddprsParameters :: Lens' DescribeDefaultParametersResponse [Parameter]

A list of parameters. Each element in the list represents one parameter.

ddprsResponseStatus :: Lens' DescribeDefaultParametersResponse Int

- | The response status code.