Network.AWS.DAX.IncreaseReplicationFactor

Description

Adds one or more nodes to a DAX cluster.

Creating a Request

increaseReplicationFactor

Arguments

:: Text
irfClusterName

-> Int
irfNewReplicationFactor

-> IncreaseReplicationFactor

Creates a value of IncreaseReplicationFactor with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

irfAvailabilityZones - The Availability Zones (AZs) in which the cluster nodes will be created. All nodes belonging to the cluster are placed in these Availability Zones. Use this parameter if you want to distribute the nodes across multiple AZs.

irfClusterName - The name of the DAX cluster that will receive additional nodes.

irfNewReplicationFactor - The new number of nodes for the DAX cluster.

data IncreaseReplicationFactor

See: increaseReplicationFactor smart constructor.

Request Lenses

irfAvailabilityZones :: Lens' IncreaseReplicationFactor [Text]

The Availability Zones (AZs) in which the cluster nodes will be created. All nodes belonging to the cluster are placed in these Availability Zones. Use this parameter if you want to distribute the nodes across multiple AZs.

irfClusterName :: Lens' IncreaseReplicationFactor Text

The name of the DAX cluster that will receive additional nodes.

irfNewReplicationFactor :: Lens' IncreaseReplicationFactor Int

The new number of nodes for the DAX cluster.

Destructuring the Response

increaseReplicationFactorResponse

Arguments

:: Int
irfrsResponseStatus

-> IncreaseReplicationFactorResponse

Creates a value of IncreaseReplicationFactorResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

irfrsCluster - A description of the DAX cluster. with its new replication factor.

irfrsResponseStatus - -- | The response status code.

data IncreaseReplicationFactorResponse

See: increaseReplicationFactorResponse smart constructor.

Response Lenses

irfrsCluster :: Lens' IncreaseReplicationFactorResponse (Maybe Cluster)

A description of the DAX cluster. with its new replication factor.

irfrsResponseStatus :: Lens' IncreaseReplicationFactorResponse Int

- | The response status code.