Network.AWS.DAX.UpdateCluster

Description

Modifies the settings for a DAX cluster. You can use this action to change one or more cluster configuration parameters by specifying the parameters and the new values.

Creating a Request

updateCluster

Arguments

:: Text
ucClusterName

-> UpdateCluster

Creates a value of UpdateCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucSecurityGroupIds - A list of user-specified security group IDs to be assigned to each node in the DAX cluster. If this parameter is not specified, DAX assigns the default VPC security group to each node.

ucPreferredMaintenanceWindow - A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00 . Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

ucNotificationTopicStatus - The current state of the topic.

ucDescription - A description of the changes being made to the cluster.

ucNotificationTopicARN - The Amazon Resource Name (ARN) that identifies the topic.

ucParameterGroupName - The name of a parameter group for this cluster.

ucClusterName - The name of the DAX cluster to be modified.

data UpdateCluster

See: updateCluster smart constructor.

Request Lenses

ucSecurityGroupIds :: Lens' UpdateCluster [Text]

A list of user-specified security group IDs to be assigned to each node in the DAX cluster. If this parameter is not specified, DAX assigns the default VPC security group to each node.

ucPreferredMaintenanceWindow :: Lens' UpdateCluster (Maybe Text)

A range of time when maintenance of DAX cluster software will be performed. For example: sun:01:00-sun:09:00 . Cluster maintenance normally takes less than 30 minutes, and is performed automatically within the maintenance window.

ucNotificationTopicStatus :: Lens' UpdateCluster (Maybe Text)

The current state of the topic.

ucDescription :: Lens' UpdateCluster (Maybe Text)

A description of the changes being made to the cluster.

ucNotificationTopicARN :: Lens' UpdateCluster (Maybe Text)

The Amazon Resource Name (ARN) that identifies the topic.

ucParameterGroupName :: Lens' UpdateCluster (Maybe Text)

The name of a parameter group for this cluster.

ucClusterName :: Lens' UpdateCluster Text

The name of the DAX cluster to be modified.

Destructuring the Response

updateClusterResponse

Arguments

:: Int
ucrsResponseStatus

-> UpdateClusterResponse

Creates a value of UpdateClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ucrsCluster - A description of the DAX cluster, after it has been modified.

ucrsResponseStatus - -- | The response status code.

data UpdateClusterResponse

See: updateClusterResponse smart constructor.

Response Lenses

ucrsCluster :: Lens' UpdateClusterResponse (Maybe Cluster)

A description of the DAX cluster, after it has been modified.

ucrsResponseStatus :: Lens' UpdateClusterResponse Int

- | The response status code.